+++ app/Http/Controllers/ManufacturerController.php
@@ -29,9 +29,12 @@ class ManufacturerController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize('view-manufacturers');
+
         $activeManufacturers = Manufacturer::count();
         $inactiveManufacturers = Manufacturer::onlyTrashed()->count();
 
@@ -50,21 +53,24 @@ class ManufacturerController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return void
      */
     public function create()
     {
-        return view('manufacturers.create');
+        // return view('manufacturers.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
+        $this->authorize('create-manufacturers');
+
         $this->service->store();
 
         session()->flash('status', 'success');
@@ -76,35 +82,46 @@ class ManufacturerController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show($id)
     {
+        $this->authorize('view-manufacturers');
+
         $manufacturer = $this->service->show($id);
+
         return view('manufacturers.show', compact('manufacturer'));
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('update-manufacturers');
+
         $manufacturer = $this->service->show($id);
+
         return view('manufacturers.edit', compact('manufacturer'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update($id)
     {
+        $this->authorize('update-manufacturers');
+
         $this->service->update($id);
 
         return redirect()->route('manufacturers.show', ['id' => $id]);
@@ -113,11 +130,14 @@ class ManufacturerController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete-manufacturers');
+
         $this->service->destroy($id);
 
         session()->flash('status', 'success');
